% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BET}
\alias{BET}
\title{Beta data}
\format{A \code{tibble} with 3 variables.
\describe{
\code{RN}  Nuclide name, e.g. Tc-99m
\code{E_MeV}  Energy grid point (MeV)
\code{A} Number of beta particles per MeV per nuclear transformation at this energy
}}
\usage{
BET
}
\description{
Beta data
}
\keyword{datasets}
