% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NDX}
\alias{NDX}
\title{Radionuclide Data Index}
\format{A \code{tibble} with 32 variables, 31 original variables from ICRP, plus the derived decay_constant variable.
\describe{
\code{RN} Name of nuclide; e.g. Am-241, Tc-99m
\code{half_life} Half-life of nuclide. Caution - units vary.
\code{units} Half-life units: ls, microsecond; ms, millisecond; s, second; m, minute; d, day; y, year
\code{decay_mode} A, alpha; B-, beta minus; B+, beta plus; EC, electron capture; IT, isomeric transition; SF, spontaneous fission
\code{pointer_rad} Not functional. Originally, location of nuclide in the ICRP-07.RAD file
\code{pointer_bet}  Not functional. Originally, location of nuclide in ICRP-07.BET file
\code{pointer_ack}  Not functional. Originally, location of nuclide in ICRP-07.ACK file
\code{pointer_nsf}  Not functional. Originally, location of nuclide in ICRP-07.NSF file
\code{progeny_1}  Radioactive progeny number 1
\code{pointer_1}  Location of progeny number 1 in RN dataset
\code{branch_1}  Branching fraction to progeny number 1
\code{progeny_2}  Radioactive progeny number 2
\code{pointer_2}  Location of progeny number 2 in RN dataset
\code{branch_2}  Branching fraction to progeny number 2
\code{progeny_3}  Radioactive progeny number 3
\code{pointer_3}  Location of progeny number 3 in RN dataset
\code{branch_3}  Branching fraction to progeny number 3
\code{progeny_4}  Radioactive progeny number 4
\code{pointer_4}  Location of progeny number 4 in RN dataset
\code{branch_}  Branching fraction to progeny number 4
\code{E_alpha}  Energy of alpha emissions (MeV/nuclear transformation)
\code{E_electron}  Energy of electrons, including beta (MeV/nuclear transformation)
\code{E_photon}  Energy of photon emission (MeV/nuclear transformation)
\code{num_phot_lt_10k}  Number of photons of energy less than 10 keV
\code{num_phot_gt_10k}  Number of photons of energy greater than 10 keV
\code{num_betas}  Number of beta transitions
\code{num_mono_e}  Number of mono-energetic electrons
\code{num_alpha}  Number of alpha transitions
\code{AMU}  Atomic mass of radionuclide (Audi et al., 2003)
\code{air_kerma_rate_const}  Air-kerma rate constant (Gy-m^2/Bq s)
\code{air_kerma_coef} Point source air-kerma coefficient (Gy-m^2/Bq s)
\code{decay_constant} Decay constant in inverse seconds.
}}
\usage{
NDX
}
\description{
The index file from ICRP-107 Nuclear Decay Data for Dosimetric Calculations
}
\keyword{datasets}
