% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RAD}
\alias{RAD}
\title{Radionuclide file}
\format{A \code{tibble} with 6 variables.
\describe{
\code{RN}  Nuclide name, e.g. Tc-99m
\code{code_AN}  Short alphanumeric code for radiation type. See rad_codes for description.
\code{E_MeV}  Numeric. Energy of the radiation in MeV
\code{prob}  Numeric. Yield of the radiation (/nuclear transformation)
\code{code_num}  Numeric code for radiation type. See rad_codes for description.
\code{is_photon} Logical. Allows selection of all photons.
}}
\usage{
RAD
}
\description{
Data on the energy and yield of each radiation emitted in nuclear transformations of the radionuclide.
}
\keyword{datasets}
