% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rad_codes}
\alias{rad_codes}
\title{Radiation code table}
\format{
A \code{tibble} with 3 variables.
\describe{
A table used for cross-linking  numerical, alphabetical and textual references for a radiation emission type, such as: 1, G, Gamma Rays.

\code{code_num}  Numeric code for radiation type. 

\code{code_AN}  Alphabetical abbreviation, some with + or -.

\code{description} Description of radiation emission type.   
}
}
\usage{
rad_codes
}
\description{
A small table providing association of \code{code_num}, \code{code_AN} and \code{description}
}
\keyword{datasets}
