% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rad_codes}
\alias{rad_codes}
\title{Radiation code table}
\format{
A \code{tibble} with 3 variables.
\describe{
\code{code_num}  Numeric code for radiation type. See rad_codes for description.
\code{code_AN}  Short alphanumeric code for radiation type. See rad_codes for description.
\code{description}  Short textual description
}
}
\usage{
rad_codes
}
\description{
A small table providing association of \code{code_num}, \code{code_AN} and \code{description}
}
\keyword{datasets}
